feat: add stricter expression placeholders
Fixes #6.

Briefly, the problem we've had all along is that there's no one
placeholder which will produce valid syntax for all positions an
expression may be interpolated into.

This can't ever really be solved perfectly as we have no idea until
run-time what the actual interpolated value will be, and what position
it is in.

Until now, we use a CSS comment as a placeholder, but this means
something like the following will not work:

```js
css`
  .foo {
    color: ${expr};
  }
`;
```

The reason is, this will become:

```css
.foo {
  color: /*PLACEHOLDER*/;
}
```

Which, from an AST point of view, is the same as having a value-less
declaration. This can cause all sorts of problems.

In this rework, we instead try to guess what kind of position we're in
based on looking ahead/behind then choosing the right placeholder
(syntax) for that position.

**This means we no longer support _all_ binding positions**. We didn't
anyway, since errors were thrown in many cases, but now we actually have
a known constraint around this.

As long as you're interpolating complete pieces of syntax in, everything
should be fine. For example, these should all be fine:

```js
css`
  ${expr} {
    color: hotpink;
  }

  .foo {
    color: ${expr};
  }

  .foo {
    ${expr}: hotpink;
  }

  .foo {
    ${expr}
  }
`;
```

While these are all near impossible for us to understand within the
linter so are no longer allowed:

```js
css`
  .foo, ${expr} {
    color: hotpink;
  }

  .foo {
    prop: ${expr}-abc;
  }
`;
```

These are all partially interpolated pieces of syntax: part of a
selector, part of a value, etc. Which will no longer work (if it ever
did).